Exception handling

+++ src/main/java/com/desafio/dscommerce/services/ProductService.java
@@ -2,15 +2,21 @@ package com.desafio.dscommerce.services;
 
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.desafio.dscommerce.DTO.ProductDTO;
 import com.desafio.dscommerce.entities.Product;
 import com.desafio.dscommerce.repositories.ProductRepository;
+import com.desafio.dscommerce.services.exceptions.DatabaseException;
 import com.desafio.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityExistsException;
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -40,15 +46,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
+        
     }
 
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+	    if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
+        try {
+            repository.deleteById(id);    		
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial.");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
